Añadida la clase GestorProducto. Falta discutir la utilidad de algunas funciones con el resto del equipo de implementacion.

+++ trunk/1_iteracion/Implementacion/SAGRES/src/GestionStock/GestionProductos/Ingrediente.java
@@ -14,7 +14,16 @@ import javax.swing.ImageIcon;
  */
 public class Ingrediente extends Producto {
 
-    public Ingrediente(int codPro, String nombre,int cantidad, int maximo, int minimo, ImageIcon imagen){
+    /**
+     * Construye un objeto Ingrediente con el estado especificado con los parametros.
+     * @param codPro Codigo identificador del ingrediente
+     * @param nombre Nombre del ingrediente
+     * @param cantidad Cantidad inicial en stock
+     * @param minimo Minimo de unidades que debe haber en stock del ingrediente
+     * @param maximo Maximo de unidades que debe haber en stock del ingrediente
+     * @param imagen Imagen del ingrediente que queremos añadir
+     */
+    public Ingrediente(int codPro, String nombre,float cantidad, float maximo, float minimo, ImageIcon imagen){
         super(imagen, nombre, minimo, maximo, cantidad, codPro);
     }
